+++ lua/gitlab/ci/context.lua
@@ -22,4 +22,26 @@ function M.from_cwd()
   return { root = root, project = project, ref = ref }, nil
 end
 
+function M.from_cwd_async(callback)
+  git.root_async(function(root, err)
+    if err then
+      callback(nil, err)
+      return
+    end
+    git.remote_project_async(function(project, err2)
+      if err2 then
+        callback(nil, err2)
+        return
+      end
+      git.branch_async(function(ref, err3)
+        if err3 then
+          callback(nil, err3)
+          return
+        end
+        callback({ root = root, project = project, ref = ref }, nil)
+      end)
+    end)
+  end)
+end
+
 return M